refactor(toolcards): dedupe toDisplayUrl by moving it to chatApi
